Store undefined symbols in `ST_user_caps`
Utilizing the new `name` field in the `cobj_symb` structure, the linker
now stores references to undefined symbols in the `ST_user_caps` array,
which itself is now inserted into the object as a kernel export.

Furthermore, the linker now also outputs the memory image of the `cobj`
header during the image output step and writes it seperately to a file
ending in `.co`.

The `pong` interface and `ping` implementation have been pulled out of
the `archive` directory for the purpose of testing this. Likewise,
additional logging information has been added to `micro_booter`.

Finally, there are a few helper comments labeled `TODO:ZD` in the
`llbooter` component in order to help guide the next step of the dynamic
loader implementation, which is mapping the object into memory.