Increased usability by cleaning up old code
I removed a number of debugging print statements that were not left intentionally.
I removed build directories related to apps that we do not and will not soon support.
The print statements in the loader should not have been left there to begin with,
but @others argues they should be kept for now because the code is not stable.
I wrapped these print statements with an #if so they are disabled by default,
and the cFE startup is not too verbose to be readable.