Implement channel validate and channel fault find
Implemented chan validate which checks that all messages on a channle (with the same id) have the same content. Implemented chan fault find which returns the replica id of the replica that published a message that differes from the majority message.

other changes: Added "unit of work" a counter on each replica that is increased each time they vote. this allows message validation to only compare messages of the same vote cycle.
moved channel code out of voter_lib and into its own file. was getting a littel big for once file.

added basic tests for above functionality to unit tests