Add task for starting 1hz timer
The `sch_lab` app appeared not to do anything because the CFE's 1hz
timer, implemented as a SIGALRM handler on POSIX, was never started.
This commit adds logic to `ostask.c` to use `sl` to serve as the timer.

Additionally, more CFE symbols have been added to the assembly stub.