Refactor Channel::channel_data
Channel::channel data was Arc<Lock<Vec<ChannelData>>> however this was completely unnecessary as the only way you could ever access the channel data was by taking the lock on the channel itself. This was removed infavor of just a Vec<ChannelData>.

This struct member was renamed from channel_data to messages. (more descriptive)