Impl voter app - rk single comp simulation
Finished the voter application api and was able to plug a demo application into the voter that does meaningless processing and attempts to make a "system call" to a service provide. that service provider is woken up when all the replicas write, and "executes the system call" then wakes the replicated app and goes back to sleep.

This is the first top to bottom "completion" (to be tested (and massively cleaned)) of the voter applicaication