Implment rep local write buffers and comp new data flag
First step in removing dynamic memory from channels. data sent by replicas is first written to a buffer stored on the replica. Once the writer passes a vote and has new data the messages are validated and combined and posted to the channel for the reader to read. This solves the memory leak issue that was occuring due to more messages being posted then there were reads when readers > writers