change symbol names for linking filesystem
During the rework of the build system, the names of a few symbols were
changed.  These were from _binary_cFEfs to _binary_cFE_fs.  This meant
that the weak symbols for the file symbols were never overwritten.
The file system did not exist so reported to the cFE that initialization
failed.

Rather than change the error in the build system changes, it was easier
to change tar.c