For migration, previously changed thd cap's size to take one full cache line.
* But I didn't fix INIT caps and I didn't test any of it in the last commit.
* per-core caps have to be on seperate cache lines, now, with thread
migration and reuse of a capability, changed thread to use a full cache
line.
* initthd, inittcap both occupy 64B cacheline.
* For the other caps, cos_kernel_api does size based and core-based
frontiers, so we're good.

* Trade off: If a thread is migrated across multiple cores, and if we
  freshly create per-core capability slot on each migration, we'd
probably be wasting a lot of space especially if we don't use that
core-local capability.
  In this simple(??) design though, we use a single capability that is as
big as a cache line and modify coreid on it to make it a core-local for
migrated core.