Fix blockpoints by complicating the usage protocol.
The race occurs as such:

1. a thread checks the data-structure, finding it needs to wait,
2. calls blkpt_wait

and thread 2 does:

a. blkpt_trigger

If preempted between 1. and 2., executing a., the wakeup will be lost.

Thus, either the meta-data tracking if there are blocked threads are internalized into the data-structure (see crt_lock.h), or the protocol above is complicated (see chan_private.h).