Fix a bug in the syscall wrapper function
We cannot use a push %rbp/%rsp because the compiler doesn't know
the inline assembly code changes the stack pointer. Thus, in some
cases the compiler might do some optimizations which could conflict
with our the push %rbp/%rsp. That means, the push might corrupt some
stack local variables. Thus, instead of using push instruction, we
use mov instruction to first save them to a stack local position(frame_ctx)
and then pass the address of the frame_ctx to the kernel rather than the sp.
When the syscall returns, the kernel will restore the sp to the address of
frame_ctx, thus we can simply use two pops to restore the original bp and sp.