Updated register protocols to simplify.
Dedicate the first three registers to pass the uniform
arguments: capability and operations for a system call, and core id,
thread token, and invocation token for upcalls.

This leaves 9 registers for general purpose use for arguments and
return values. This enables standardizing the API for argument access
by special-casing the dedicated registers vs. the general ones.